refactor: BOOLEAN and connect()
- refactor: merge connect() with Realm
- refactor: tinyint(1) to boolean
  - MySQL uses TINYINT(1) directly
  - PostgreSQL maps BOOLEAN to int
  - SQLite maps boolean to tinyint
- fix: findOne in sequelize adapter
- fix: value masking in logger.logQuery()
