+++ lib/adapters/sequelize.js
@@ -133,7 +133,7 @@ module.exports = Bone => {
         // findOne(id)
         return super.findOne(options);
       }
-      return this.findAll({ ...options, limit: 1 });
+      return this.findAll({ ...options, limit: 1 }).later(result => result[0]);
     }
 
     static async findOrBuild(options = {}) {
