+++ lib/bone.js
@@ -37,7 +37,6 @@ function compare(attributes, columns) {
     const { columnName } = definition;
 
     if (!definition.equals(columnMap[columnName])) {
-      console.log(definition, columnMap[columnName])
       diff[name] = {
         modify: columnMap.hasOwnProperty(columnName),
         ...definition,
@@ -581,7 +580,7 @@ class Bone {
         return value ? JSON.parse(value) : null;
       case Date:
         return value instanceof Date ? value : new Date(value);
-      // node-sqlite3 doesn't handle TINYINT(1) as BOOLEAN well
+      // node-sqlite3 doesn't convert TINYINT(1) to boolean by default
       case Boolean:
         return Boolean(value);
       default:
