+++ lib/data_types.js
@@ -110,12 +110,11 @@ class DATE extends DataType {
 class BOOLEAN extends DataType {
   constructor() {
     super();
-    this.dataType = 'tinyint';
+    this.dataType = 'boolean';
   }
 
   toSqlString() {
-    const dataType = this.dataType.toUpperCase();
-    return `${dataType}(1)`;
+    return this.dataType.toUpperCase();
   }
 }
