+++ lib/drivers/abstract/definition.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const debug = require('debug')('leoric');
+
 const DataType = require('../../data_types');
 const { snakeCase } = require('../../utils/string');
 
@@ -56,7 +58,13 @@ class Definition {
     if (!targetDefinition) return false;
     const props = [ 'dataType', 'allowNull', 'defaultValue', 'primaryKey' ];
     for (const prop of props) {
-      if (!Object.is(this[prop], targetDefinition[prop])) {
+      // SQLite has default value as string even if data type is integer
+      if (this[prop] != targetDefinition[prop]) {
+        debug('[definition] %s [%s] not equal (defined: %s, actual: %s)',
+          this.columnName,
+          prop,
+          this[prop],
+          targetDefinition[prop]);
         return false;
       }
     }
