+++ lib/drivers/abstract/logger.js
@@ -11,15 +11,15 @@ class Logger {
 
   format(query, values, opts = {}) {
     const { command, sets } = opts;
+
     if ([ 'insert', 'upsert', 'update' ].includes(command) && sets) {
       const { hideKeys } = this;
       const keys = Object.keys(sets);
-      for (let i = 0; i < keys.length; i++) {
-        const key = keys[i];
-        if (hideKeys.includes(key)) {
-          values[i] = '***';
-        }
-      }
+      values = values.map((entry, i) => {
+        // INSERT ... ON DUPLICATE UPDATE ...
+        const index = command === 'upsert' ? i % keys.length : i;
+        return hideKeys.includes(keys[index]) ? '***' : entry;
+      });
     }
     return SqlString.format(query.sql || query, values);
   }
