+++ lib/drivers/mysql/definition.js
@@ -4,6 +4,14 @@ const SqlString = require('sqlstring');
 const Definition = require('../abstract/definition');
 
 class MysqlDefinition extends Definition {
+  constructor(name, params, opts) {
+    super(name, params, opts);
+    if (this.dataType === 'boolean') {
+      this.dataType = 'tinyint';
+      this.columnType = 'TINYINT(1)';
+    }
+  }
+
   toSqlString() {
     const {
       type, columnType, dataType, allowNull, defaultValue, primaryKey,
