+++ lib/drivers/sqlite/definition.js
@@ -3,6 +3,11 @@
 const SqlString = require('sqlstring');
 const Definition = require('../abstract/definition');
 
+function escape(value) {
+  if (typeof value === 'boolean') return +value;
+  return SqlString.escape(value);
+}
+
 class SqliteDefinition extends Definition {
   constructor(name, params, opts) {
     super(name, params, opts);
@@ -24,7 +29,7 @@ class SqliteDefinition extends Definition {
     if (!primaryKey && !allowNull) chunks.push('NOT NULL');
 
     if (defaultValue != null) {
-      chunks.push(`DEFAULT ${SqlString.escape(defaultValue)}`);
+      chunks.push(`DEFAULT ${escape(defaultValue)}`);
     }
 
     return chunks.join(' ');
