+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "leoric",
-  "version": "0.5.0-alpha.8",
+  "version": "0.5.0-alpha.15",
   "description": "JavaScript Object-relational mapping alchemy",
   "main": "index.js",
   "types": "index.d.ts",
